+++ android/app/src/main/java/ir/zitar/planner/MainActivity.java
@@ -1,4 +1,4 @@
-package ir.sepidar.productivity;
+package ir.zitar.planner;
 
 import com.getcapacitor.BridgeActivity;
 import android.os.Bundle;
@@ -7,7 +7,8 @@ public class MainActivity extends BridgeActivity {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         registerPlugin(DocumentExportPlugin.class);
-        registerPlugin(SepidarWidgetPlugin.class);
+        registerPlugin(ZitarWidgetPlugin.class);
+        registerPlugin(ZitarCalendarPlugin.class);
         super.onCreate(savedInstanceState);
     }
 }