+++ capacitor.config.ts
@@ -1,13 +1,13 @@
 import type { CapacitorConfig } from '@capacitor/cli'
 
 const config: CapacitorConfig = {
-  appId: 'ir.sepidar.productivity',
-  appName: 'سپیدار',
+  appId: 'ir.zitar.planner',
+  appName: 'زیتر',
   webDir: 'dist',
   bundledWebRuntime: false,
   plugins: {
     LocalNotifications: {
-      smallIcon: 'ic_stat_sepidar',
+      smallIcon: 'ic_stat_zitar',
       iconColor: '#4773FA'
     }
   }